requirements.txt
aniso8601==7.0.0
asgiref==3.5.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
cryptography==37.0.2
Django==3.2.15
django-admin-ip-restrictor==2.2.0
django-cors-headers==3.11.0
django-filter==22.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-ipware==3.0.7
djangorestframework==3.13.1
drf-jwt==1.19.2
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
msgpack==1.0.4
promise==2.3
psycopg2==2.9.3
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.11
uWSGI==2.0.20
